Fix: NavBar 안 읽은 메시지 배지 숫자 반영 오류 해결
